Fixed a small comparison warning (signed vs. unsigned)
git-svn-id: http://source.sphene.net/svn/xfirelib/trunk@115 9c1b9215-2909-0410-98e2-e07619621686

+++ xfirelib/src/testclient/test.cpp
@@ -192,7 +192,7 @@ namespace xfirelibtest {
   }
   string XFireTestClient::joinString(vector<string> s, int startindex, int endindex, string delimiter) {
     string ret;
-    int max = endindex > s.size() || endindex < 0 ? s.size() : endindex;
+    int max = (unsigned)endindex > s.size() || endindex < 0 ? s.size() : endindex;
     for(int i = startindex ; i < max ; i++) {
       if(i != startindex)
 	ret += delimiter;
